new routes for top tracks and deatiled stats working
